Placed the component did mount and state in the incorrect file. Moved them to the correct one. Everything is working as intended still. Added some CSS as well
