+++ src/components/Character.css
@@ -1,0 +1,15 @@
+
+.character-container {
+    font-size: 1.5rem;
+    margin-left: 5%;
+    border: 1px solid red;
+    border-radius: 50%;
+    background: black;
+    color: white;
+    padding: 4% 11%;
+    line-height: 3.2rem;
+}
+
+.character-container li {
+    list-style-type: none;
+}
