+++ src/components/CharacterList.css
@@ -1,0 +1,8 @@
+ul {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+    margin-top: 12%;
+
+}
