+++ src/views/CharacterListView.js
@@ -2,20 +2,19 @@ import React from "react";
 import { connect } from "react-redux";
 
 import { CharacterList } from "../components";
+import { getChars } from '../actions';
 // import actions
 
 class CharacterListView extends React.Component {
-  constructor() {
-    super();
-  }
 
-  componentDidMount() {
-    // call our action
+  componentDidMount(){
+    this.props.getChars();
   }
 
   render() {
     if (this.props.fetching) {
       // return something here to indicate that you are fetching data
+      return <div>Loading...</div>
     }
     return (
       <div className="CharactersList_wrapper">
@@ -27,9 +26,14 @@ class CharacterListView extends React.Component {
 
 // our mapStateToProps needs to have two properties inherited from state
 // the characters and the fetching boolean
+
+const mapStateToProps = state => ({
+  characters: state.charsReducer.characters,
+  error: state.charsReducer.error,
+  fetchingChars: state.charsReducer.fetchingChars
+});
+
 export default connect(
-  null /* mapStateToProps replaces null here */,
-  {
-    /* action creators go here */
-  }
+  mapStateToProps,
+  { getChars }
 )(CharacterListView);
